refactor(threads): useQueuedSend 外移纯 helper 区至 queuedSendHelpers.ts
- 2648→2154 行；56-549 纯区逐字外移（drain signal/slash command/fuse 判定等）
- drain flag 单例随 helpers 走，主文件 re-export 锁测试面（77 测试零改动通过）
- UseQueuedSendOptions/Result 契约类型留主文件
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；useQueuedSend.test.tsx 77 passed；gate 绿
