+++ docs/architecture/large-file-new-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T16:15:47.699Z",
+  "generatedAt": "2026-08-30T16:24:11.568Z",
   "scope": "new-file",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
@@ -584,16 +584,6 @@
       "thresholdSource": "new-file-ratchet"
     },
     {
-      "path": "src/features/threads/hooks/useQueuedSend.ts",
-      "lines": 2648,
-      "type": "ts/tsx",
-      "policyId": "feature-hotpath",
-      "priority": "P1",
-      "warnThreshold": 2400,
-      "failThreshold": 800,
-      "thresholdSource": "new-file-ratchet"
-    },
-    {
       "path": "src/features/files/components/FileTreePanel.tsx",
       "lines": 2622,
       "type": "ts/tsx",
@@ -826,6 +816,16 @@
     {
       "path": "src/features/threads/hooks/useThreadActionsResumeThread.ts",
       "lines": 2293,
+      "type": "ts/tsx",
+      "policyId": "feature-hotpath",
+      "priority": "P1",
+      "warnThreshold": 2400,
+      "failThreshold": 800,
+      "thresholdSource": "new-file-ratchet"
+    },
+    {
+      "path": "src/features/threads/hooks/useQueuedSend.ts",
+      "lines": 2287,
       "type": "ts/tsx",
       "policyId": "feature-hotpath",
       "priority": "P1",
